pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64", "setuptools-scm>=8"]

[project]
authors = [{ name = "Conrad Bzura", email = "conradbzura@gmail.com" }]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = ["aiohttp", "click", "fastapi", "globus-sdk>=4.1.0", "motor", "pymongo", "strawberry-graphql", "uvicorn", "wool"]
description = "A Python utility for parsing and normalizing various DCC datapackages."
dynamic = ["version"]
license = { text = "MIT License" }
name = "dcc2cvh"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["debugpy", "ruff"]

[project.scripts]
dcc2cvh = "dcc2cvh.cli:cli"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
pythonpath = "src"

[tool.setuptools_scm]
local_scheme = "dirty-tag"
version_scheme = "release-branch-semver"
